requirements.txt
django>=3.2.4,<5.0.1
sqlparse==0.4.1
mysqlclient==2.2